move CachableCallback -> hub.api
